refactor(trending): move under own folder
