+++ projects/client/src/routes/movies/+page.svelte
@@ -7,7 +7,7 @@
   import AnticipatedList from "$lib/sections/lists/anticipated/AnticipatedList.svelte";
   import PopularList from "$lib/sections/lists/popular/PopularList.svelte";
   import RecommendedList from "$lib/sections/lists/RecommendedList.svelte";
-  import TrendingList from "$lib/sections/lists/TrendingList.svelte";
+  import TrendingList from "$lib/sections/lists/trending/TrendingList.svelte";
   import {
     DEFAULT_COVER,
     DEFAULT_SHARE_MOVIE_COVER,
