+++ projects/client/src/routes/movies/trending/+page.svelte
@@ -2,7 +2,7 @@
   import TraktPage from "$lib/components/layout/TraktPage.svelte";
   import * as m from "$lib/features/i18n/messages";
 
-  import TrendingPaginatedList from "$lib/sections/lists/TrendingPaginatedList.svelte";
+  import TrendingPaginatedList from "$lib/sections/lists/trending/TrendingPaginatedList.svelte";
   import { DEFAULT_SHARE_MOVIE_COVER } from "$lib/utils/constants";
 </script>
